Implement: export methods from Queries.scala to ElasticIntervalQuery.scala
